👌 IMPROVE: Improve course filter and useQueryString hook
