+++ src/components/filter/index.jsx
@@ -1,3 +1,2 @@
 export { FilterAside, FilterDropdown } from './FilterContainer'
-export { default as FilterItem } from './FilterItem'
 export { default as FilterFloatButton } from './FilterFloatButton'
